requirements.txt
appdirs==1.4.3
attrs==19.3.0
beautifulsoup4==4.9.0
black==19.10b0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cssmin==0.2.0
docutils==0.16
feedgenerator==1.9.1
idna==2.9
Jinja2==2.11.3
Markdown==3.1.1
MarkupSafe==1.1.1
pathspec==0.7.0
pelican==4.6.0
pelican-webassets==2.0.0
Pygments==2.6.1
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2021.1
regex==2020.2.20
requests==2.23.0
six==1.14.0
soupsieve==2.0
toml==0.10.0
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.8
webassets==2.0
